.github/script/requirements.txt
requests>=2.22.0
newsapi-python>=0.2.6